agent-admin: reorganize urls, views and tests (#774)
