+++ ami/agent_admin/tests/test_home.py
@@ -9,7 +9,7 @@ def test_home_with_support_agent(app, support_agent: Agent) -> None:
     app.set_user(support_agent.user)
     response = app.get("/agent-admin/", status=200)
     assert "/agent-admin/manage/access/" not in response
-    assert "/agent-admin/notification/" not in response
+    assert "/agent-admin/manage/notification/" not in response
 
 
 @pytest.mark.django_db
@@ -17,7 +17,7 @@ def test_home_with_notifications_agent(app, notifications_agent: Agent) -> None:
     app.set_user(notifications_agent.user)
     response = app.get("/agent-admin/", status=200)
     assert "/agent-admin/manage/access/" not in response
-    assert "/agent-admin/notification/" in response
+    assert "/agent-admin/manage/notification/" in response
 
 
 @pytest.mark.django_db
@@ -25,7 +25,7 @@ def test_home_with_admin_agent(app, admin_agent: Agent) -> None:
     app.set_user(admin_agent.user)
     response = app.get("/agent-admin/", status=200)
     assert "/agent-admin/manage/access/" in response
-    assert "/agent-admin/notification/" in response
+    assert "/agent-admin/manage/notification/" in response
 
 
 @pytest.mark.django_db
