--- ami/agent_admin/views.py
@@ -1,140 +1,0 @@
-from django.shortcuts import redirect, render
-from django.urls import reverse
-
-from ami.agent.decorators import (
-    agent_login_required,
-    role_admin_required,
-    role_notifications_required,
-    role_support_required,
-)
-from ami.agent.models import Agent
-from ami.agent_admin.forms import AgentFormSet, NotificationForm
-from ami.agent_admin.models import AuditEntry
-
-
-@agent_login_required
-@role_support_required
-def home(request):
-    return render(request, "agent_admin/home.html", {})
-
-
-def login(request):
-    return render(request, "agent_admin/login.html", {})
-
-
-@agent_login_required
-def access_denied(request):
-    home_url = reverse("agent-admin:home")
-    context = {
-        "btn_group": {
-            "items": [
-                {
-                    "label": "Recharger",
-                    "type": "button",
-                    "extra_classes": "fr-btn--secondary",
-                    "onclick": f"window.location.href = '{home_url}';",
-                },
-            ],
-            "extra_classes": "fr-btns-group--inline fr-btns-group--form-actions",
-        },
-    }
-    return render(request, "agent_admin/access_denied.html", context)
-
-
-@agent_login_required
-@role_admin_required
-def manage_access(request):
-    unauthorized_agents = Agent.objects.filter(role__isnull=True).order_by("-proconnect_last_login")
-    authorized_agents = Agent.objects.filter(role__isnull=False).order_by(
-        "user__last_name",
-        "user__first_name",
-    )
-
-    if request.method == "POST":
-        unauthorized_agents_formset = AgentFormSet(
-            request.POST,
-            queryset=unauthorized_agents,
-            prefix="unauthorized",
-            form_kwargs={"author": request.user.agent},
-        )
-        authorized_agents_formset = AgentFormSet(
-            request.POST,
-            queryset=authorized_agents,
-            prefix="authorized",
-            form_kwargs={"author": request.user.agent},
-        )
-        is_unauthorized_agents_formset_valid = unauthorized_agents_formset.is_valid()
-        is_authorized_agents_formset_valid = authorized_agents_formset.is_valid()
-        if is_unauthorized_agents_formset_valid and is_authorized_agents_formset_valid:
-            unauthorized_agents_formset.save()
-            authorized_agents_formset.save()
-            return redirect(reverse("agent-admin:manage-access"))
-    else:
-        unauthorized_agents_formset = AgentFormSet(
-            queryset=unauthorized_agents,
-            prefix="unauthorized",
-            form_kwargs={"author": request.user.agent},
-        )
-        authorized_agents_formset = AgentFormSet(
-            queryset=authorized_agents,
-            prefix="authorized",
-            form_kwargs={"author": request.user.agent},
-        )
-
-    context = {
-        "unauthorized_agents_formset": unauthorized_agents_formset,
-        "authorized_agents_formset": authorized_agents_formset,
-        "btn_group": {
-            "items": [
-                {
-                    "label": "Annuler",
-                    "type": "button",
-                    "extra_classes": "fr-btn--secondary",
-                    "onclick": "window.location.href = window.location.href;",
-                },
-                {
-                    "label": "Enregistrer",
-                    "type": "submit",
-                },
-            ],
-            "extra_classes": "fr-btns-group--inline fr-btns-group--form-actions",
-        },
-        "aes": AuditEntry.objects.filter(
-            action_type="access",
-            action_code__in=["role-added", "role-updated", "role-removed"],
-        ),
-    }
-
-    return render(request, "agent_admin/manage_access.html", context)
-
-
-@agent_login_required
-@role_notifications_required
-def send_notification(request):
-    result_data_dict = None
-    if request.method == "POST":
-        form = NotificationForm(data=request.POST)
-        if form.is_valid():
-            result_data_dict = form.submit()
-    else:
-        form = NotificationForm()
-    context = {
-        "form": form,
-        "form_result_data_dict": result_data_dict,
-        "btn_group": {
-            "items": [
-                {
-                    "label": "Annuler",
-                    "type": "button",
-                    "extra_classes": "fr-btn--secondary",
-                    "onclick": "window.location.href = window.location.href;",
-                },
-                {
-                    "label": "Enregistrer",
-                    "type": "submit",
-                },
-            ],
-            "extra_classes": "fr-btns-group--inline fr-btns-group--form-actions",
-        },
-    }
-    return render(request, "agent_admin/send_notification.html", context)
